ml/requirements.txt
pandas~=2.0.2
scikit-learn~=1.2.2
nltk~=3.8.1
joblib
liac-arff